pyproject.toml
[tool.poetry]
name = "sophia"
version = "0.1.0"
description = ""
authors = ["taha <mokfi.taha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
transformers = "^4.44.2"
google-cloud-storage = "^2.18.2"
librosa = "^0.10.2.post1"
ffmpeg-python = "^0.2.0"
llama-index-core = "^0.11.11"
llama-index-llms-openai = "^0.2.9"
llama-index = "^0.11.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
